Move `failing` tests to `tofix`
