+++ blackbird/src/test/java/com/fasterxml/jackson/module/blackbird/tofix/TestClassloaders.java
@@ -1,4 +1,4 @@
-package com.fasterxml.jackson.module.blackbird.failing;
+package com.fasterxml.jackson.module.blackbird.tofix;
 
 import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import com.fasterxml.jackson.module.blackbird.BlackbirdTestBase;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.module.blackbird.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -19,6 +20,7 @@ public class TestClassloaders extends BlackbirdTestBase
 
     // Note: this test always passes in Java 8, even if the issue is not fixed,
     // so it is duplicated in jackson-jdk11-compat-test for now
+    @JacksonTestFailureExpected
     @Test
     public void testLoadInChildClassloader() throws Exception {
         TestLoader loader = new TestLoader(getClass().getClassLoader());
