+++ jaxb/src/test/java/com/fasterxml/jackson/module/jaxb/tofix/TestEnums256.java
@@ -1,4 +1,4 @@
-package com.fasterxml.jackson.module.jaxb.failing;
+package com.fasterxml.jackson.module.jaxb.tofix;
 
 import javax.xml.bind.annotation.*;
 import javax.xml.bind.annotation.adapters.*;
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.module.jaxb.BaseJaxbTest;
+import com.fasterxml.jackson.module.jaxb.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -44,6 +45,7 @@ public class TestEnums256 extends BaseJaxbTest
     private final ObjectMapper MAPPER = getJaxbMapper();
 
     // [modules-base#256]
+    @JacksonTestFailureExpected
     @Test
     public void testEnumSerialize256() throws Exception
     {
