+++ pom.xml
@@ -115,11 +115,6 @@ not datatype, data format, or JAX-RS provider modules.
 	<plugin>
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-surefire-plugin</artifactId>
-          <configuration>
-            <excludes>
-              <exclude>com/fasterxml/jackson/**/failing/*.java</exclude>
-            </excludes>
-          </configuration>
 	</plugin>
       </plugins>
     </pluginManagement>
